Add option to disable web security
Electron is strict with regards to cross-domain requests.
If Content-Type is application/json, then it sends a preflight
requests which should be handled by server properly. This might not
be preferrable for the tested web server that is configured in production
mode with CORS disabled.

Cross-domain requests are useful for controling environment of system
under test i.e. setting DBus properties.